requirements.txt
# Core dependencies
dspy-ai>=2.4.0
openai>=1.0.0
pandas>=2.0.0
openpyxl>=3.1.0

# Utilities
python-dotenv>=1.0.0
